backend/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
supabase
pydantic
email-validator